requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.2
djangorestframework==3.12.1
ipykernel==5.3.4
jupyter-client==6.1.6
pytz==2020.1
pyzmq==19.0.2
sqlparse==0.4.1
beautifulsoup4==4.8.0
whitenoise==3.3.1
gunicorn==20.0.4
psycopg2==2.7.5
whitenoise==5.1.0